pyproject.toml
[tool.poetry]
name = "insanely-fast-whisper-api"
version = "0.2.0"
description = "An insanely fast whisper API"
authors = ["Yoeven D Khemlani <yoevenkhem@gmail.com>"]
packages = [{ include = "app" }]
readme = "README.md"
license = "MIT"


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
gunicorn = "^21.2.0"
uvicorn = { extras = ["standard"], version = "^0.27.1" }
transformers = "^4.37.2"
optimum = "^1.17.0"
accelerate = "^0.27.2"
setuptools = "^69.1.0"
pyannote-audio = "^3.1.1"
pydantic = "^2.6.1"



[tool.pdm]
[tool.pdm.build]
includes = ["app"]

[tool.pdm.dev-dependencies]
dev = [
    "uvicorn>=0.32.0",
]
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
authors = [
    {name = "Yoeven D Khemlani", email = "yoevenkhem@gmail.com"},
]
license = {text = "MIT"}
requires-python = "<4.0,>=3.10"
dependencies = [
    "fastapi<1.0.0,>=0.109.2",
    "gunicorn<22.0.0,>=21.2.0",
    "uvicorn[standard]<1.0.0,>=0.27.1",
    "transformers<5.0.0,>=4.37.2",
    "optimum<2.0.0,>=1.17.0",
    "accelerate<1.0.0,>=0.27.2",
    "setuptools<70.0.0,>=69.1.0",
    "pyannote-audio<4.0.0,>=3.1.1",
    "pydantic<3.0.0,>=2.6.1",
]
name = "insanely-fast-whisper-api"
version = "0.2.0"
description = "An insanely fast whisper API"
readme = "README.md"
